move frame configuration at the task's toplevel
Unless I am mistaken, the body frame and output's source frame
are required to be the same. Moreover, the convention for frame
configuration (as done in the transformer) is to have a property
called <framename>_frame, e.g. here world_frame. This convention
is required by e.g. Syskit for the transformer configuration